[KAFKAWEBSITE-17] add migration file, rename FLywayService
