+++ src/main/java/ru/kfu/kafkaWebSite/service/datasource/FlywayExtension.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import ru.kfu.kafkaWebSite.service.FlywayService;
 
 public class FlywayExtension implements BeforeAllCallback {
